requirements/base.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
psycopg2-binary>=2.9.5
django-environ>=0.10.0
django-cors-headers>=3.14.0
celery>=5.2.7
redis>=4.5.4
channels>=4.0.0
channels-redis>=4.1.0
drf-spectacular>=0.26.0
django-filter>=23.1
Pillow>=9.5.0
